feat:sinario n+2 et fixer tous les problemes

+++ src/main/java/com/atb/employeeevaluation/dto/EvaluationN1Request.java
@@ -14,7 +14,7 @@ public class EvaluationN1Request {
     private Long evaluationId;
 
     @NotNull(message = "Les réponses sont obligatoires")
-    private Map<Long, Integer> reponses;  // Question ID -> Note
+    private Map<Long, Double> reponses;  // Question ID -> Note
 
     private String commentaire;
 }
